feat(symfony): make the symfony container actually work
