+++ .gitignore
@@ -1,3 +1,4 @@
 /vendor/
 /coverage/
+/build/
 composer.lock
