+++ src/MockeryContainer/PsrMockeryContainerDecorator.php
@@ -3,12 +3,16 @@ declare(strict_types = 1);
 
 namespace MockeryContainer;
 
+use Mockery\MockInterface;
 use Psr\Container\ContainerInterface;
 use Psr\Container\NotFoundExceptionInterface;
 
 class PsrMockeryContainerDecorator implements ContainerInterface, MockeryContainerInterface
 {
-    use MockeryContainerTrait;
+    /**
+     * @var MockeryContainerInterface
+     */
+    private $mockeryContainer;
 
     /**
      * @var ContainerInterface
@@ -45,5 +49,15 @@ class PsrMockeryContainerDecorator implements ContainerInterface, MockeryContain
         }
 
         return $this->container->has($id);
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function mock(string $id, ...$args): MockInterface
+    {
+        array_unshift($args, $id);
+
+        return call_user_func_array([$this->mockeryContainer, 'mock'], $args);
     }
 }
